// backend/src/rpc/types/gen/PtTALLiNtPec/Block.ts
export type Block = BlockInfoEncodingV1;
export type _024_PtTALLiN$ContractId = string;
// ...
export type Operation =
  | {
      protocol: "PtTALLiNtPec7mE7yY4m3k26J8Qukef3E3ehzhfXgFZKGtDdAXu";
      chain_id: _024_PtTALLiN$ContractId;
      hash: _024_PtTALLiN$ContractId;
      branch: _024_PtTALLiN$ContractId;
      contents: _024_PtTALLiN$Operation$Alpha$Contents[];
      signature?: _024_PtTALLiN$ContractId;
      metadata: "too large";
    }
  | {
      protocol: "PtTALLiNtPec7mE7yY4m3k26J8Qukef3E3ehzhfXgFZKGtDdAXu";
      chain_id: _024_PtTALLiN$ContractId;
      hash: _024_PtTALLiN$ContractId;
      branch: _024_PtTALLiN$ContractId;
      contents: _024_PtTALLiN$Operation$Alpha$Contents[];
      signature?: _024_PtTALLiN$ContractId;
    }
  | {
      protocol: "PtTALLiNtPec7mE7yY4m3k26J8Qukef3E3ehzhfXgFZKGtDdAXu";
      chain_id: _024_PtTALLiN$ContractId;
      hash: _024_PtTALLiN$ContractId;
      branch: _024_PtTALLiN$ContractId;
      contents: _024_PtTALLiN$Operation$Alpha$OperationContentsAndResult[];
      signature?: _024_PtTALLiN$ContractId;
    }
  | {
      protocol: "PtTALLiNtPec7mE7yY4m3k26J8Qukef3E3ehzhfXgFZKGtDdAXu";
      chain_id: _024_PtTALLiN$ContractId;
      hash: _024_PtTALLiN$ContractId;
      branch: _024_PtTALLiN$ContractId;
      contents: _024_PtTALLiN$Operation$Alpha$Contents[];
      signature?: _024_PtTALLiN$ContractId;
    };
export type _024_PtTALLiN$Operation$Alpha$Contents =
  | Attestation
  | AttestationWithDal
  | Preattestation
  | AttestationsAggregate
  | PreattestationsAggregate
  | DoubleConsensusOperationEvidence
  | SeedNonceRevelation
  | VdfRevelation
  | DoubleBakingEvidence
  | DalEntrapmentEvidence
  | ActivateAccount
  | Proposals
  | Ballot
  | Reveal1
  | Transaction
  | Origination1
  | Delegation1
  | SetDepositsLimit1
  | IncreasePaidStorage1
  | UpdateConsensusKey1
  | UpdateCompanionKey
  | DrainDelegate
  | FailingNoop
  | RegisterGlobalConstant
  | TransferTicket
  | DalPublishCommitment
  | SmartRollupOriginate1
  | SmartRollupAddMessages
  | SmartRollupCement
  | SmartRollupPublish
  | SmartRollupRefute
  | SmartRollupTimeout
  | SmartRollupExecuteOutboxMessage
  | SmartRollupRecoverBond
  | ZkRollupOrigination
  | ZkRollupPublish
  | ZkRollupUpdate;
// ...
export type _024_PtTALLiN$Operation$Alpha$OperationContentsAndResult =
  | Attestation2
  | AttestationWithDal2
  | Preattestation2
  | AttestationsAggregate2
  | PreattestationsAggregate2
  | DoubleConsensusOperationEvidence1
  | DalEntrapmentEvidence1
  | SeedNonceRevelation1
  | VdfRevelation1
  | DoubleBakingEvidence1
  | ActivateAccount1
  | Proposals1
  | Ballot1
  | DrainDelegate1
  | Reveal2
  | Transaction2
  | Origination3
  | Delegation3
  | RegisterGlobalConstant1
  | SetDepositsLimit2
  | IncreasePaidStorage2
  | UpdateConsensusKey2
  | UpdateCompanionKey1
  | TransferTicket1
  | DalPublishCommitment1
  | SmartRollupOriginate2
  | SmartRollupAddMessages1
  | SmartRollupCement1
  | SmartRollupPublish1
  | SmartRollupRefute1
  | SmartRollupTimeout1
  | SmartRollupExecuteOutboxMessage1
  | SmartRollupRecoverBond1
  | ZkRollupOrigination1
  | ZkRollupPublish1
  | ZkRollupUpdate1;
// ...
export interface BlockInfoEncodingV1 {
  protocol: "PtTALLiNtPec7mE7yY4m3k26J8Qukef3E3ehzhfXgFZKGtDdAXu";
  chain_id: _024_PtTALLiN$ContractId;
  hash: _024_PtTALLiN$ContractId;
  header: ShellHeader;
  metadata?: BlockHeaderMetadata;
  operations: Operation[][];
}
/**
 * Block header's shell-related content. It contains information such as the block level, its predecessor and timestamp.
 */
export interface ShellHeader {
  level: number;
  proto: number;
  predecessor: _024_PtTALLiN$ContractId;
  timestamp: _024_PtTALLiN$ContractId;
  validation_pass: number;
  operations_hash: _024_PtTALLiN$ContractId;
  fitness: BlockFitness;
  context: _024_PtTALLiN$ContractId;
  payload_hash: _024_PtTALLiN$ContractId;
  payload_round: number;
  proof_of_work_nonce: string;
  seed_nonce_hash?: _024_PtTALLiN$ContractId;
  liquidity_baking_toggle_vote: _024_PtTALLiN$LiquidityBakingVote;
  signature: _024_PtTALLiN$ContractId;
}
export interface BlockHeaderMetadata {
  protocol: "PtTALLiNtPec7mE7yY4m3k26J8Qukef3E3ehzhfXgFZKGtDdAXu";
  next_protocol: "PtTALLiNtPec7mE7yY4m3k26J8Qukef3E3ehzhfXgFZKGtDdAXu";
  test_chain_status: TestChainStatus;
  max_operations_ttl: number;
  max_operation_data_length: number;
  max_block_header_length: number;
  max_operation_list_length: {
    max_size: number;
    max_op?: number;
  }[];
  proposer: _024_PtTALLiN$ContractId;
  baker: _024_PtTALLiN$ContractId;
  level_info: {
    /**
     * The level of the block relative to genesis. This is also the Shell's notion of level.
     */
    level: number;
    /**
     * The level of the block relative to the successor of the genesis block. More precisely, it is the position of the block relative to the block that starts the "Alpha family" of protocols, which includes all protocols except Genesis (that is, from 001 onwards).
     */
    level_position: number;
    /**
     * The current cycle's number. Note that cycles are a protocol-specific notion. As a result, the cycle number starts at 0 with the first block of the Alpha family of protocols.
     */
    cycle: number;
    /**
     * The current level of the block relative to the first block of the current cycle.
     */
    cycle_position: number;
    /**
     * Tells whether the baker of this block has to commit a seed nonce hash.
     */
    expected_commitment: boolean;
  };
  voting_period_info: {
    /**
     * The voting period to which the block belongs.
     */
    voting_period: {
      /**
       * The voting period's index. Starts at 0 with the first block of the Alpha family of protocols.
       */
      index: number;
      kind: Proposal | Exploration | Cooldown | Promotion | Adoption;
      /**
       * The relative position of the first level of the period with respect to the first level of the Alpha family of protocols.
       */
      start_position: number;
    };
    /**
     * The position of the block within the voting period.
     */
    position: number;
    /**
     * The number of blocks remaining till the end of the voting period.
     */
    remaining: number;
  };
  nonce_hash: _024_PtTALLiN$ContractId | None;
  deactivated: _024_PtTALLiN$ContractId[];
  balance_updates?: _024_PtTALLiN$OperationMetadata$Alpha$BalanceUpdates;
  liquidity_baking_toggle_ema: number;
  implicit_operations_results: _024_PtTALLiN$Operation$Alpha$SuccessfulManagerOperationResult[];
  proposer_consensus_key: _024_PtTALLiN$ContractId;
  baker_consensus_key: _024_PtTALLiN$ContractId;
  consumed_milligas: PositiveBigNumber;
  dal_attestation: BigNumber;
  all_bakers_attest_activation_level: number | null;
  attestations: { total_committee_power: string; threshold: string; recorded_power: string } | null;
  preattestations: { total_committee_power: string; threshold: string; recorded_power: string } | null;
}
// ...
export interface Attestation {
  kind: "attestation";
  slot: number;
  level: number;
  round: number;
  block_payload_hash: _024_PtTALLiN$ContractId;
}
export interface AttestationWithDal {
  kind: "attestation_with_dal";
  slot: number;
  level: number;
  round: number;
  block_payload_hash: _024_PtTALLiN$ContractId;
  dal_attestation: BigNumber;
}
export interface Preattestation {
  kind: "preattestation";
  slot: number;
  level: number;
  round: number;
  block_payload_hash: _024_PtTALLiN$ContractId;
}
export interface AttestationsAggregate {
  kind: "attestations_aggregate";
  consensus_content: {
    level: number;
    round: number;
    block_payload_hash: _024_PtTALLiN$ContractId;
  };
  committee: {
    slot: number;
    dal_attestation?: BigNumber;
  }[];
}
export interface PreattestationsAggregate {
  kind: "preattestations_aggregate";
  consensus_content: {
    level: number;
    round: number;
    block_payload_hash: _024_PtTALLiN$ContractId;
  };
  committee: number[];
}
export interface DoubleConsensusOperationEvidence {
  kind: "double_consensus_operation_evidence";
  slot: number;
  op1: _024_PtTALLiN$Inlined$ConsensusOperation;
  op2: _024_PtTALLiN$Inlined$ConsensusOperation;
}
// ...
export interface SeedNonceRevelation {
  kind: "seed_nonce_revelation";
  level: number;
  nonce: string;
}
export interface VdfRevelation {
  kind: "vdf_revelation";
  solution: [] | [string] | [string, string];
}
export interface DoubleBakingEvidence {
  kind: "double_baking_evidence";
  bh1: ShellHeader1;
  bh2: ShellHeader1;
}
// ...
export interface DalEntrapmentEvidence {
  kind: "dal_entrapment_evidence";
  attestation: _024_PtTALLiN$Inlined$ConsensusOperation;
  consensus_slot: number;
  slot_index: number;
  shard_with_proof: {
    shard: [] | [number] | [number, string[]];
    proof: _024_PtTALLiN$ContractId;
  };
}
export interface ActivateAccount {
  kind: "activate_account";
  pkh: _024_PtTALLiN$ContractId;
  secret: string;
}
export interface Proposals {
  kind: "proposals";
  source: _024_PtTALLiN$ContractId;
  period: number;
  proposals: _024_PtTALLiN$ContractId[];
}
export interface Ballot {
  kind: "ballot";
  source: _024_PtTALLiN$ContractId;
  period: number;
  proposal: _024_PtTALLiN$ContractId;
  ballot: "nay" | "yay" | "pass";
}
export interface Reveal1 {
  kind: "reveal";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  public_key: _024_PtTALLiN$ContractId;
  proof?: _024_PtTALLiN$ContractId;
}
export interface Transaction {
  kind: "transaction";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  amount: PositiveBigNumber;
  destination: _024_PtTALLiN$ContractId;
  parameters?: {
    entrypoint: _024_PtTALLiN$Entrypoint;
    value: unknown;
  };
}
export interface Origination1 {
  kind: "origination";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  balance: PositiveBigNumber;
  delegate?: _024_PtTALLiN$ContractId;
  script: _024_PtTALLiN$Scripted$Contracts;
}
export interface _024_PtTALLiN$Scripted$Contracts {
  code: unknown;
  storage: unknown;
}
export interface Delegation1 {
  kind: "delegation";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  delegate?: _024_PtTALLiN$ContractId;
}
export interface SetDepositsLimit1 {
  kind: "set_deposits_limit";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  limit?: PositiveBigNumber;
}
export interface IncreasePaidStorage1 {
  kind: "increase_paid_storage";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  amount: BigNumber;
  destination: _024_PtTALLiN$ContractId;
}
export interface UpdateConsensusKey1 {
  kind: "update_consensus_key";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  pk: _024_PtTALLiN$ContractId;
  proof?: _024_PtTALLiN$ContractId;
}
export interface UpdateCompanionKey {
  kind: "update_companion_key";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  pk: _024_PtTALLiN$ContractId;
  proof?: _024_PtTALLiN$ContractId;
}
export interface DrainDelegate {
  kind: "drain_delegate";
  consensus_key: _024_PtTALLiN$ContractId;
  delegate: _024_PtTALLiN$ContractId;
  destination: _024_PtTALLiN$ContractId;
}
export interface FailingNoop {
  kind: "failing_noop";
  arbitrary: string;
}
export interface RegisterGlobalConstant {
  kind: "register_global_constant";
  source: _024_PtTALLiN$ContractId;
  fee: PositiveBigNumber;
  counter: PositiveBigNumber;
  gas_limit: PositiveBigNumber;
  storage_limit: PositiveBigNumber;
  value: unknown;
}